+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.21-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.4.21-alpha.0...@selfcommunity/react-core@0.4.21-alpha.1) (2024-05-27)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
+## [0.4.21-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.4.20...@selfcommunity/react-core@0.4.21-alpha.0) (2024-05-24)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
 ## [0.4.20](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.4.20-alpha.2...@selfcommunity/react-core@0.4.20) (2024-05-22)
 
 **Note:** Version bump only for package @selfcommunity/react-core