+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.21-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@0.2.20...@selfcommunity/utils@0.2.21-alpha.0) (2024-05-24)
+
+**Note:** Version bump only for package @selfcommunity/utils
+
+
+
+
+
 ## [0.2.20](https://github.com/selfcommunity/community-js/compare/@selfcommunity/utils@0.2.20-alpha.1...@selfcommunity/utils@0.2.20) (2024-05-22)
 
 **Note:** Version bump only for package @selfcommunity/utils